Added file ownership

+++ src/main/java/edu/sigmaportal/platform/controller/FilesController.java
@@ -1,7 +1,12 @@
 package edu.sigmaportal.platform.controller;
 
 import edu.sigmaportal.platform.dto.FileDto;
+import edu.sigmaportal.platform.exception.EntityNotFoundException;
+import edu.sigmaportal.platform.exception.InsufficientPermissionsException;
+import edu.sigmaportal.platform.service.CourseService;
+import edu.sigmaportal.platform.service.EnrollmentsService;
 import edu.sigmaportal.platform.service.FileService;
+import edu.sigmaportal.platform.util.AuthUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -12,12 +17,15 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.annotation.Secured;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.UncheckedIOException;
+import java.util.Optional;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
@@ -28,9 +36,13 @@ import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
 public class FilesController {
 
     private final FileService service;
+    private final EnrollmentsService enrolls;
+    private final CourseService courses;
 
-    public FilesController(FileService service) {
+    public FilesController(FileService service, EnrollmentsService enrolls, CourseService courses) {
         this.service = service;
+        this.enrolls = enrolls;
+        this.courses = courses;
     }
 
     @GetMapping(value = "/{id}", produces = APPLICATION_JSON_VALUE)
@@ -39,8 +51,20 @@ public class FilesController {
             @ApiResponse(responseCode = "200", description = "File found", content = @Content(schema = @Schema(implementation = FileDto.class))),
             @ApiResponse(responseCode = "404", description = "File not found", content = @Content)
     })
-    public FileDto findById(@Parameter(description = "ID of the file") @PathVariable String id) {
-        return service.find(id);
+    public FileDto findById(@Parameter(description = "ID of the file") @PathVariable String id, Authentication auth) {
+        String userId = AuthUtils.getUserId(auth);
+        if (service.owns(userId, id))
+            return service.find(id);
+
+        Optional<String> courseId = service.connectedCourseId(id);
+        if (courseId.isEmpty())
+            throw new EntityNotFoundException("File not found");
+
+        if (enrolls.isEnrolled(courseId.get(), userId) || courses.owns(userId, courseId.get())) {
+            return service.find(id);
+        }
+
+        throw new EntityNotFoundException("File not found");
     }
 
     @GetMapping("/{id}/download/")
@@ -49,16 +73,25 @@ public class FilesController {
             @ApiResponse(responseCode = "200", description = "File bytes sent"),
             @ApiResponse(responseCode = "404", description = "File not found")
     })
-    public void download(@Parameter(description = "ID of the file") @PathVariable String id, HttpServletResponse response) {
-        FileService.FileStream stream = service.stream(id);
-        response.setContentType(stream.mimeType());
-        response.setHeader("Content-Disposition", "attachment; filename=\"" + stream.filename() + "\"");
-        try (InputStream s = stream.handle()) {
-            s.transferTo(response.getOutputStream());
-            response.flushBuffer();
-        } catch (IOException e) {
-            throw new UncheckedIOException(e);
+    public void download(@Parameter(description = "ID of the file") @PathVariable String id, HttpServletResponse response, Authentication auth) {
+        Optional<String> courseId = service.connectedCourseId(id);
+        if (courseId.isEmpty())
+            throw new EntityNotFoundException("File not found");
+
+        String userId = AuthUtils.getUserId(auth);
+        if (enrolls.isEnrolled(courseId.get(), userId) || courses.owns(userId, courseId.get())) {
+            FileService.FileStream stream = service.stream(id);
+            response.setContentType(stream.mimeType());
+            response.setHeader("Content-Disposition", "attachment; filename=\"" + stream.filename() + "\"");
+            try (InputStream s = stream.handle()) {
+                s.transferTo(response.getOutputStream());
+                response.flushBuffer();
+            } catch (IOException e) {
+                throw new UncheckedIOException(e);
+            }
         }
+
+        throw new EntityNotFoundException("File not found");
     }
 
     @PostMapping(value = "/", produces = APPLICATION_JSON_VALUE, consumes = APPLICATION_JSON_VALUE)
@@ -67,8 +100,10 @@ public class FilesController {
             @ApiResponse(responseCode = "200", description = "File created", content = @Content(schema = @Schema(implementation = FileDto.class))),
             @ApiResponse(responseCode = "400", description = "Invalid file object", content = @Content)
     })
-    public FileDto create(@Valid @RequestBody FileDto file) {
-        return service.upload(file);
+    @Secured("create_file")
+    public FileDto create(@Valid @RequestBody FileDto file, Authentication auth) {
+        String userId = AuthUtils.getUserId(auth);
+        return service.upload(userId, file);
     }
 
     @PostMapping(value = "/upload", consumes = MULTIPART_FORM_DATA_VALUE, produces = APPLICATION_JSON_VALUE)
@@ -77,8 +112,10 @@ public class FilesController {
             @ApiResponse(responseCode = "200", description = "File created", content = @Content(schema = @Schema(implementation = FileDto.class))),
             @ApiResponse(responseCode = "400", description = "Invalid file object", content = @Content)
     })
-    public FileDto upload(@RequestParam("file") MultipartFile file) {
-        return service.upload(file);
+    @Secured("create_file")
+    public FileDto upload(@RequestParam("file") MultipartFile file, Authentication auth) {
+        String userId = AuthUtils.getUserId(auth);
+        return service.upload(userId, file);
     }
 
     @PatchMapping(value = "/{id}", consumes = APPLICATION_JSON_VALUE, produces = APPLICATION_JSON_VALUE)
@@ -88,8 +125,14 @@ public class FilesController {
             @ApiResponse(responseCode = "400", description = "Invalid file object", content = @Content),
             @ApiResponse(responseCode = "404", description = "File not found", content = @Content)
     })
-    public FileDto update(@Parameter(description = "ID of the file") @PathVariable String id, @RequestBody FileDto file) {
-        return service.update(id, file);
+    @Secured("edit_file")
+    public FileDto update(@Parameter(description = "ID of the file") @PathVariable String id, @RequestBody FileDto file, Authentication auth) {
+        String userId = AuthUtils.getUserId(auth);
+        if (service.owns(userId, id)) {
+            return service.update(id, file);
+        }
+
+        throw new InsufficientPermissionsException("Not a file owner");
     }
 
     @DeleteMapping("/{id}")
@@ -98,8 +141,14 @@ public class FilesController {
             @ApiResponse(responseCode = "204", description = "File deleted"),
             @ApiResponse(responseCode = "404", description = "File not found")
     })
-    public ResponseEntity<Void> delete(@Parameter(description = "ID of the file") @PathVariable String id) {
-        service.delete(id);
-        return ResponseEntity.noContent().build();
+    @Secured("delete_file")
+    public ResponseEntity<Void> delete(@Parameter(description = "ID of the file") @PathVariable String id, Authentication auth) {
+        String userId = AuthUtils.getUserId(auth);
+        if (service.owns(userId, id)) {
+            service.delete(id);
+            return ResponseEntity.noContent().build();
+        }
+
+        throw new InsufficientPermissionsException("Not a file owner");
     }
 }
